--- a/pkg/apis/monitoring/v1/prometheus_types.go
+++ b/pkg/apis/monitoring/v1/prometheus_types.go
@@ -1829,5 +1829,5 @@ type ScrapeClass struct {
 	// More info: https://prometheus.io/docs/prometheus/latest/configuration/configuration/#relabel_config
 	//
 	// +optional
-	Relabelings []*RelabelConfig `json:"relabelings,omitempty"`
+	Relabelings []RelabelConfig `json:"relabelings,omitempty"`
 }