--- a/pkg/client/applyconfiguration/monitoring/v1/scrapeclass.go
+++ b/pkg/client/applyconfiguration/monitoring/v1/scrapeclass.go
@@ -16,17 +16,13 @@
 
 package v1
 
-import (
-	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
-)
-
 // ScrapeClassApplyConfiguration represents an declarative configuration of the ScrapeClass type for use
 // with apply.
 type ScrapeClassApplyConfiguration struct {
-	Name        *string                       `json:"name,omitempty"`
-	Default     *bool                         `json:"default,omitempty"`
-	TLSConfig   *TLSConfigApplyConfiguration  `json:"tlsConfig,omitempty"`
-	Relabelings []*monitoringv1.RelabelConfig `json:"relabelings,omitempty"`
+	Name        *string                           `json:"name,omitempty"`
+	Default     *bool                             `json:"default,omitempty"`
+	TLSConfig   *TLSConfigApplyConfiguration      `json:"tlsConfig,omitempty"`
+	Relabelings []RelabelConfigApplyConfiguration `json:"relabelings,omitempty"`
 }
 
 // ScrapeClassApplyConfiguration constructs an declarative configuration of the ScrapeClass type for use with
@@ -62,7 +58,7 @@ func (b *ScrapeClassApplyConfiguration) WithTLSConfig(value *TLSConfigApplyConfi
 // WithRelabelings adds the given value to the Relabelings field in the declarative configuration
 // and returns the receiver, so that objects can be build by chaining "With" function invocations.
 // If called multiple times, values provided by each call will be appended to the Relabelings field.
-func (b *ScrapeClassApplyConfiguration) WithRelabelings(values ...**monitoringv1.RelabelConfig) *ScrapeClassApplyConfiguration {
+func (b *ScrapeClassApplyConfiguration) WithRelabelings(values ...*RelabelConfigApplyConfiguration) *ScrapeClassApplyConfiguration {
 	for i := range values {
 		if values[i] == nil {
 			panic("nil value passed to WithRelabelings")