--- a/pkg/prometheus/resource_selector.go
+++ b/pkg/prometheus/resource_selector.go
@@ -247,13 +247,14 @@ func validateScrapeIntervalAndTimeout(p monitoringv1.PrometheusInterface, scrape
 	return CompareScrapeTimeoutToScrapeInterval(scrapeTimeout, scrapeInterval)
 }
 
-func validateRelabelConfigs(p monitoringv1.PrometheusInterface, rcs []*monitoringv1.RelabelConfig) error {
+func validateRelabelConfigs(p monitoringv1.PrometheusInterface, rcs []monitoringv1.RelabelConfig) error {
 	for i, rc := range rcs {
-		if rc == nil {
-			return fmt.Errorf("null relabel config")
-		}
+		// No need for this now
+		//	if rc == nil {
+		//		return fmt.Errorf("null relabel config")
+		//	}
 
-		if err := validateRelabelConfig(p, *rc); err != nil {
+		if err := validateRelabelConfig(p, rc); err != nil {
 			return fmt.Errorf("[%d]: %w", i, err)
 		}
 	}