… pkg/client/applyconfiguration/monitoring/v1alpha1/scrapeconfigspec.go
… pkg/prometheus/promcfg_test.go
… pkg/prometheus/resource_selector_test.go