## Description
Relabel configs previously were of type `[]*monitoringv1.RelabelConfigs`, this PR changes that to `[]monitoringv1.RelabelConfigs`. 
Earlier users could write invalid manifests like this one:
```
relabelConfigs:
- null
- sourceLabels: ...
```

Which the operator would then have to validate and then reject. By changing the type from pointer to just the value, we are able to reject such invalid manifests directly without having to reject it at the operator level.

This PR is to check the viability of this change. 
Related discussion: https://kubernetes.slack.com/archives/C01B03QCSMN/p1712306476986909, [#6379 (comment)](https://github.com/prometheus-operator/prometheus-operator/pull/6379#discussion_r1521132751)

## Type of change

_What type of changes does your code introduce to the Prometheus operator? Put an `x` in the box that apply._

- [ ] `CHANGE` (fix or feature that would cause existing functionality to not work as expected)
- [ ] `FEATURE` (non-breaking change which adds functionality)
- [ ] `BUGFIX` (non-breaking change which fixes an issue)
- [ ] `ENHANCEMENT` (non-breaking change which improves existing functionality)
- [x] `NONE` (if none of the other choices apply. Example, tooling, build system, CI, docs, etc.)

## Verification
<!-- How you tested it? How do you know it works? -->
Please check the [Prometheus-Operator testing guidelines](../TESTING.md) for recommendations about automated tests.

## Changelog entry

_Please put a one-line changelog entry below. This will be copied to the changelog file during the release process._

<!-- 
Your release note should be written in clear and straightforward sentences. Most often, users aren't familiar with
the technical details of your PR, so consider what they need to know when you write your release note.

Some brief examples of release notes:
- Add metadataConfig field to the Prometheus CRD for configuring how remote-write sends metadata information.
- Generate correct scraping configuration for Probes with empty or unset module parameter.
-->

```
Change type of the different RelabelConfigs fields to `[]monitoringv1.RelabelConfigs` instead of an array of pointers
```