@ArthurSens 2024-04-12
I just did the benchmark myself:

<details><summary>Benchmark code</summary>

```go
func BenchmarkSelectServiceMonitors(b *testing.B) {
	ca, err := os.ReadFile("../../test/e2e/remote_write_certs/ca.crt")
	require.NoError(b, err)

	cert, err := os.ReadFile("../../test/e2e/remote_write_certs/client.crt")
	require.NoError(b, err)

	key, err := os.ReadFile("../../test/e2e/remote_write_certs/client.key")
	require.NoError(b, err)

	cs := fake.NewSimpleClientset(
		&v1.Secret{
			ObjectMeta: metav1.ObjectMeta{
				Name:      "secret",
				Namespace: "test",
			},
			Data: map[string][]byte{
				"ca":         ca,
				"invalid_ca": []byte("garbage"),
				"cert":       cert,
				"key":        key,
			},
		},
		&v1.ConfigMap{
			ObjectMeta: metav1.ObjectMeta{
				Name:      "configmap",
				Namespace: "test",
			},
			Data: map[string]string{
				"ca":   string(ca),
				"cert": string(cert),
			},
		},
	)

	rs := NewResourceSelector(
		newLogger(),
		&monitoringv1.Prometheus{
			Spec: monitoringv1.PrometheusSpec{
				CommonPrometheusFields: monitoringv1.CommonPrometheusFields{},
			},
		},
		assets.NewStore(cs.CoreV1(), cs.CoreV1()),
		nil,
		operator.NewMetrics(prometheus.NewPedanticRegistry()),
		record.NewFakeRecorder(1),
	)

	// Some common K8s relabeling configs
	rcs := make([]*monitoringv1.RelabelConfig, 15)
	for i := range rcs {
		rcs[i] = &monitoringv1.RelabelConfig{
			SourceLabels: []monitoringv1.LabelName{"__name__"},
			Action:       "drop",
			Regex:        "kubelet_(pod_worker_latency_microseconds|pod_start_latency_microseconds|cgroup_manager_latency_microseconds|pod_worker_start_latency_microseconds|pleg_relist_latency_microseconds|pleg_relist_interval_microseconds|runtime_operations|runtime_operations_latency_microseconds|runtime_operations_errors|eviction_stats_age_microseconds|device_plugin_registration_count|device_plugin_alloc_latency_microseconds|network_plugin_operations_latency_microseconds)",
		}
	}

	sm := &monitoringv1.ServiceMonitor{
		ObjectMeta: metav1.ObjectMeta{
			Name:      "test",
			Namespace: "test",
		},
		Spec: monitoringv1.ServiceMonitorSpec{
			Endpoints: []monitoringv1.Endpoint{
				{
					RelabelConfigs: rcs,
				},
			},
		},
	}

	b.ResetTimer()

	for i := 0; i < b.N; i++ {
		_, err := rs.SelectServiceMonitors(context.Background(), func(_ string, _ labels.Selector, appendFn cache.AppendFunc) error {
			appendFn(sm)
			return nil
		})
		require.NoError(b, err)
	}
}

```

</details>

It appears what I said above is incorrect or irrelevant when compared to everything else used in SelecServiceMonitors :)

Results:
```console
goos: darwin
goarch: arm64
pkg: github.com/prometheus-operator/prometheus-operator/pkg/prometheus
                         │ ../prometheus-operator/bench-with-pointers.txt │ ../prometheus-operator/bench-without-pointers.txt │
                         │                     sec/op                     │             sec/op               vs base          │
SelectServiceMonitors-12                                      653.3µ ± 4%                       655.4µ ± 2%  ~ (p=0.684 n=10)

                         │ ../prometheus-operator/bench-with-pointers.txt │ ../prometheus-operator/bench-without-pointers.txt │
                         │                      B/op                      │            B/op             vs base               │
SelectServiceMonitors-12                                     1.638Mi ± 0%                 1.638Mi ± 0%  -0.00% (p=0.000 n=10)

                         │ ../prometheus-operator/bench-with-pointers.txt │ ../prometheus-operator/bench-without-pointers.txt │
                         │                   allocs/op                    │         allocs/op           vs base               │
SelectServiceMonitors-12                                      4.408k ± 0%                  4.393k ± 0%  -0.34% (p=0.000 n=10)

```